separated many Screen classes to its modules
