+++ src/ksigner.kv
@@ -122,15 +122,14 @@
 		rows: 5
 
 		Button:
+			id: verify_screen_load_file_button
 			text: 'Load file'			
 			font_size: root.height // 25
         	font_name: 'terminus.ttf'
 			background_color: 0, 0, 0, 0,
 			color: 1, 1, 1, 1
-			on_press:
-				self.background_color = (.5, .5, .5, .5)
-    		on_release:
-				self.background_color = (0, 0, 0, 0)
+			on_press: root.on_press_verify_screen_load_file_button()
+    		on_release: root.on_release_verify_screen_load_file_button()
 			canvas.before:
         		Color:
             		rgba: 1, 1, 1, 1
